@@ -21,16 +21,9 @@
 
 namespace ddn\sapp;
 
-use ddn\sapp\PDFObject;
-use ddn\sapp\pdfvalue\PDFValue;
-use ddn\sapp\pdfvalue\PDFValueHexString;
-use ddn\sapp\pdfvalue\PDFValueList;
-use ddn\sapp\pdfvalue\PDFValueObject;
-use ddn\sapp\pdfvalue\PDFValueReference;
+use function ddn\sapp\helpers\timestamp_to_pdfdatestring;
 use ddn\sapp\pdfvalue\PDFValueSimple;
 use ddn\sapp\pdfvalue\PDFValueString;
-use ddn\sapp\pdfvalue\PDFValueType;
-use function ddn\sapp\helpers\timestamp_to_pdfdatestring;
 
 // This is an special object that has a set of fields
 class PDFSignatureObject extends PDFObject {
@@ -42,27 +35,34 @@ class PDFSignatureObject extends PDFObject {
     // is not known. 68 digits enable 20 digits for the size of the document
     public static $__BYTERANGE_SIZE = 68;
 
-    protected $_prev_content_size = 0;
+    protected int $_prev_content_size;
+
     protected $_post_content_size = null;
 
     // A placeholder for the certificate to use to sign the document
     protected $_certificate = null;
+
     protected $_signature_ltv_data = null;
+
     protected $_signature_tsa = null;
+
     /**
      * Sets the certificate to use to sign
      * @param cert the pem-formatted certificate and private to use to sign as
      *             [ 'cert' => ..., 'pkey' => ... ]
      */
-    public function set_certificate($certificate) {
+    public function set_certificate($certificate): void {
         $this->_certificate = $certificate;
     }
-    public function set_signature_ltv($signature_ltv_data) {
+
+    public function set_signature_ltv($signature_ltv_data): void {
         $this->_signature_ltv_data = $signature_ltv_data;
     }
-    public function set_signature_tsa($signature_tsa) {
+
+    public function set_signature_tsa($signature_tsa): void {
         $this->_signature_tsa = $signature_tsa;
     }
+
     /**
      * Obtains the certificate set with function set_certificate
      * @return cert the certificate
@@ -70,12 +70,15 @@ class PDFSignatureObject extends PDFObject {
     public function get_certificate() {
         return $this->_certificate;
     }
+
     public function get_tsa() {
         return $this->_signature_tsa;
     }
+
     public function get_ltv() {
         return $this->_signature_ltv_data;
     }
+
     /**
      * Constructs the object and sets the default values needed to sign
      * @param oid the oid for the object
@@ -92,6 +95,7 @@ class PDFSignatureObject extends PDFObject {
             'M' => new PDFValueString(timestamp_to_pdfdatestring()),
         ]);
     }
+
     /**
      * Function used to add some metadata fields to the signature: name, reason of signature, etc.
      * @param name the name of the signer
@@ -99,7 +103,7 @@ class PDFSignatureObject extends PDFObject {
      * @param location the location of signature
      * @param contact the contact info
      */
-    public function set_metadata($name = null, $reason = null, $location = null, $contact = null) {
+    public function set_metadata($name = null, $reason = null, $location = null, $contact = null): void {
         if ($name !== null) {
             $this->_value["Name"] = new PDFValueString($name);
         }
@@ -113,46 +117,50 @@ class PDFSignatureObject extends PDFObject {
             $this->_value["ContactInfo"] = new PDFValueString($contact);
         }
     }
+
     /**
      * Function that sets the size of the content that will appear in the file, previous to this object,
      *   and the content that will be included after. This is needed to get the range of bytes of the
      *   signature.
      */
-    public function set_sizes($prev_content_size, $post_content_size = null) {
+    public function set_sizes(int $prev_content_size, $post_content_size = null): void {
         $this->_prev_content_size = $prev_content_size;
         $this->_post_content_size = $post_content_size;
     }
+
     /**
      * This function gets the offset of the marker, relative to this object. To make correct, the offset of the object
      *   shall have properly been set. It makes use of the parent "to_pdf_entry" function to avoid recursivity.
      * @return position the position of the <0000 marker
      */
-    public function get_signature_marker_offset() {
+    public function get_signature_marker_offset(): int {
         $tmp_output = parent::to_pdf_entry();
         $marker = "/Contents";
         $position = strpos($tmp_output, $marker);
         return $position + strlen($marker);
     }
+
     /**
      * Overrides the parent function to calculate the proper range of bytes, according to the sizes provided and the
      *   string representation of this object
      * @return str the string representation of this object
      */
-    public function to_pdf_entry() {
+    public function to_pdf_entry(): string {
         $signature_size = strlen(parent::to_pdf_entry());
         $offset = $this->get_signature_marker_offset();
         $starting_second_part = $this->_prev_content_size + $offset + self::$__SIGNATURE_MAX_LENGTH + 2;
 
         $contents_size = strlen("" . $this->_value['Contents']);
 
-        $byterange_str =  "[ 0 " .
+        $byterange_str = "[ 0 " .
             ($this->_prev_content_size + $offset) . " " .
             ($starting_second_part) . " " .
-            ($this->_post_content_size!==null?$this->_post_content_size + ($signature_size - $contents_size - $offset):0) . " ]";
+            ($this->_post_content_size !== null ? $this->_post_content_size + ($signature_size - $contents_size - $offset) : 0) . " ]";
 
         $this->_value['ByteRange'] =
-            new PDFValueSimple($byterange_str . str_repeat(" ", self::$__BYTERANGE_SIZE - strlen($byterange_str) + 1)
-        );
+            new PDFValueSimple(
+                $byterange_str . str_repeat(" ", self::$__BYTERANGE_SIZE - strlen($byterange_str) + 1)
+            );
 
         return parent::to_pdf_entry();
     }
