@@ -21,44 +21,48 @@
 
 namespace ddn\sapp\helpers;
 
-if (!defined('__CONVENIENT_MAX_BUFFER_DUMP'))
-    define('__CONVENIENT_MAX_BUFFER_DUMP', 80);
+use Stringable;
 
-use function ddn\sapp\helpers\debug_var;
+if (! defined('__CONVENIENT_MAX_BUFFER_DUMP'))
+    define('__CONVENIENT_MAX_BUFFER_DUMP', 80);
 
 /**
  * This class is used to manage a buffer of characters. The main features are that
  *   it is possible to add data (by usign *data* function), and getting the current
  *   size. Then it is possible to get the whole buffer using function *get_raw*
  */
-class Buffer {
+class Buffer implements Stringable {
     protected $_buffer = "";
-    protected $_bufferlen = 0;
+
+    protected int $_bufferlen;
 
     public function __construct($string = null) {
         if ($string === null)
             $string = "";
 
         $this->_buffer = $string;
-        $this->_bufferlen = strlen($string);
+        $this->_bufferlen = strlen((string) $string);
     }
+
     /**
      * Adds raw data to the buffer
      * @param data the data to add
      */
-    public function data(...$datas) {
+    public function data(...$datas): void {
         foreach ($datas as $data) {
-            $this->_bufferlen += strlen($data);
+            $this->_bufferlen += strlen((string) $data);
             $this->_buffer .= $data;
         }
-    }    
+    }
+
     /**
      * Obtains the size of the buffer
      * @return size the size of the buffer
      */
-    public function size() {
+    public function size(): int {
         return $this->_bufferlen;
     }
+
     /**
      * Gets the raw data from the buffer
      * @return buffer the raw data
@@ -66,27 +70,29 @@ class Buffer {
     public function get_raw() {
         return $this->_buffer;
     }
+
     /**
      * Appends buffer $b to this buffer
      * @param b the buffer to be added to this one
      * @return buffer this object
      */
-    public function append($b) {
-        if (get_class($b) !== get_class($this))
+    public function append($b): static {
+        if ($b::class !== static::class)
             throw new Exception('invalid buffer to add to this one');
-        
+
         $this->_buffer .= $b->get_raw();
         $this->_bufferlen = strlen($this->_buffer);
         return $this;
     }
+
     /**
      * Obtains a new buffer that is the result from the concatenation of this buffer and the parameter
      * @param b the buffer to be added to this one
      * @return buffer the resulting buffer (different from this one)
      */
-    public function add(...$bs) {
+    public function add(...$bs): \ddn\sapp\helpers\Buffer {
         foreach ($bs as $b) {
-            if (get_class($b) !== get_class($this))
+            if ($b::class !== static::class)
                 throw new Exception('invalid buffer to add to this one');
         }
 
@@ -96,33 +102,34 @@ class Buffer {
 
         return $r;
     }
+
     /**
      * Returns a new buffer that contains the same data than this one
      * @return buffer the cloned buffer
      */
-    public function clone() {
+    public function clone(): \ddn\sapp\helpers\Buffer {
         $buffer = new Buffer($this->_buffer);
         return $buffer;
     }
+
     /**
      * Provides a easy to read string representation of the buffer, using the "var_dump" output
      *   of the variable, but providing a reduced otput of the buffer
      * @return str a string with the representation of the buffer
      */
-    public function __toString() {
-        if (strlen($this->_buffer) < (__CONVENIENT_MAX_BUFFER_DUMP * 2))
-            return debug_var($this);
+    public function __toString(): string {
+        if (strlen((string) $this->_buffer) < (__CONVENIENT_MAX_BUFFER_DUMP * 2))
+            return (string) debug_var($this);
 
         $buffer = $this->_buffer;
-        $this->_buffer = substr($buffer, 0, __CONVENIENT_MAX_BUFFER_DUMP);
-        $this->_buffer .= "\n...\n" . substr($buffer, -__CONVENIENT_MAX_BUFFER_DUMP);
+        $this->_buffer = substr((string) $buffer, 0, __CONVENIENT_MAX_BUFFER_DUMP);
+        $this->_buffer .= "\n...\n" . substr((string) $buffer, -__CONVENIENT_MAX_BUFFER_DUMP);
         $result = debug_var($this);
         $this->_buffer = $buffer;
-        return $result;
+        return (string) $result;
     }    
 
-
-    public function show_bytes($columns, $offset = 0, $length = null) {
+    public function show_bytes($columns, $offset = 0, $length = null): string {
         if ($length === null)
             $length = $this->_bufferlen;
 
