@@ -1,4 +1,5 @@
 <?php
+
 namespace ddn\sapp\helpers;
 /*
 // File name   : asn1.php
@@ -19,35 +20,35 @@ class asn1 {
    * @return string asn.1 tag name
    */
   protected static function type($id) {
-    $asn1_Types = array(
-      "00" => "ASN1_EOC",
-      "01" => "ASN1_BOOLEAN",
-      "02" => "ASN1_INTEGER",
-      "03" => "ASN1_BIT_STRING",
-      "04" => "ASN1_OCTET_STRING",
-      "05" => "ASN1_NULL",
-      "06" => "ASN1_OBJECT",
-      "07" => "ASN1_OBJECT_DESCRIPTOR",
-      "08" => "ASN1_EXTERNAL",
-      "09" => "ASN1_REAL",
-      "0a" => "ASN1_ENUMERATED",
-      "0c" => "ASN1_UTF8STRING",
-      "30" => "ASN1_SEQUENCE",
-      "31" => "ASN1_SET",
-      "12" => "ASN1_NUMERICSTRING",
-      "13" => "ASN1_PRINTABLESTRING",
-      "14" => "ASN1_T61STRING",
-      "15" => "ASN1_VIDEOTEXSTRING",
-      "16" => "ASN1_IA5STRING",
-      "17" => "ASN1_UTCTIME",
-      "18" => "ASN1_GENERALIZEDTIME",
-      "19" => "ASN1_GRAPHICSTRING",
-      "1a" => "ASN1_VISIBLESTRING",
-      "1b" => "ASN1_GENERALSTRING",
-      "1c" => "ASN1_UNIVERSALSTRING",
-      "1d" => "ASN1_BMPSTRING"
-    );
-    return array_key_exists($id,$asn1_Types)?$asn1_Types[$id]:$id;
+    $asn1_Types = [
+        "00" => "ASN1_EOC",
+        "01" => "ASN1_BOOLEAN",
+        "02" => "ASN1_INTEGER",
+        "03" => "ASN1_BIT_STRING",
+        "04" => "ASN1_OCTET_STRING",
+        "05" => "ASN1_NULL",
+        "06" => "ASN1_OBJECT",
+        "07" => "ASN1_OBJECT_DESCRIPTOR",
+        "08" => "ASN1_EXTERNAL",
+        "09" => "ASN1_REAL",
+        "0a" => "ASN1_ENUMERATED",
+        "0c" => "ASN1_UTF8STRING",
+        "30" => "ASN1_SEQUENCE",
+        "31" => "ASN1_SET",
+        "12" => "ASN1_NUMERICSTRING",
+        "13" => "ASN1_PRINTABLESTRING",
+        "14" => "ASN1_T61STRING",
+        "15" => "ASN1_VIDEOTEXSTRING",
+        "16" => "ASN1_IA5STRING",
+        "17" => "ASN1_UTCTIME",
+        "18" => "ASN1_GENERALIZEDTIME",
+        "19" => "ASN1_GRAPHICSTRING",
+        "1a" => "ASN1_VISIBLESTRING",
+        "1b" => "ASN1_GENERALSTRING",
+        "1c" => "ASN1_UNIVERSALSTRING",
+        "1d" => "ASN1_BMPSTRING",
+    ];
+    return array_key_exists($id, $asn1_Types) ? $asn1_Types[$id] : $id;
   }
 
   /**
@@ -60,7 +61,7 @@ class asn1 {
     if($hex == '') {
       return false;
     }
-    if(!@ctype_xdigit($hex) || @strlen($hex)%2!=0) {
+    if(! @ctype_xdigit($hex) || @strlen($hex) % 2 != 0) {
       echo "input:\"$hex\" not hex string!.\n";
       return false;
     }
@@ -69,29 +70,29 @@ class asn1 {
       $asn1_type = substr($hex, 0, 2);
       $tlv_tagLength = hexdec(substr($hex, 2, 2));
       if($tlv_tagLength > 127) {
-        $tlv_lengthLength = $tlv_tagLength-128;
-        $tlv_valueLength = substr($hex, 4, ($tlv_lengthLength*2));
+        $tlv_lengthLength = $tlv_tagLength - 128;
+        $tlv_valueLength = substr($hex, 4, ($tlv_lengthLength * 2));
       } else {
         $tlv_lengthLength = 0;
-        $tlv_valueLength = substr($hex, 2, 2+($tlv_lengthLength*2));
+        $tlv_valueLength = substr($hex, 2, 2 + ($tlv_lengthLength * 2));
       }
-      if($tlv_lengthLength >4) { // limit tlv_lengthLength to FFFF
+      if($tlv_lengthLength > 4) { // limit tlv_lengthLength to FFFF
         return false;
       }
       $tlv_valueLength = hexdec($tlv_valueLength);
-      $totalTlLength = 2+2+($tlv_lengthLength*2);
-      $reduction = 2+2+($tlv_lengthLength*2)+($tlv_valueLength*2);
-      $tlv_value = substr($hex, $totalTlLength, $tlv_valueLength*2);
-      $remain = substr($hex, $totalTlLength+($tlv_valueLength*2));
-      $newhexdump = substr($hex, 0, $totalTlLength+($tlv_valueLength*2));
-      $result[] = array(
-        'tlv_tagLength'=>strlen(dechex($tlv_tagLength))%2==0?dechex($tlv_tagLength):'0'.dechex($tlv_tagLength),
-        'tlv_lengthLength'=>$tlv_lengthLength,
-        'tlv_valueLength'=>$tlv_valueLength,
-        'newhexdump'=>$newhexdump,
-        'typ'=>$asn1_type,
-        'tlv_value'=>$tlv_value
-      );
+      $totalTlLength = 2 + 2 + ($tlv_lengthLength * 2);
+      $reduction = 2 + 2 + ($tlv_lengthLength * 2) + ($tlv_valueLength * 2);
+      $tlv_value = substr($hex, $totalTlLength, $tlv_valueLength * 2);
+      $remain = substr($hex, $totalTlLength + ($tlv_valueLength * 2));
+      $newhexdump = substr($hex, 0, $totalTlLength + ($tlv_valueLength * 2));
+      $result[] = [
+        'tlv_tagLength' => strlen(dechex($tlv_tagLength)) % 2 == 0 ? dechex($tlv_tagLength) : '0' . dechex($tlv_tagLength),
+          'tlv_lengthLength' => $tlv_lengthLength,
+          'tlv_valueLength' => $tlv_valueLength,
+          'newhexdump' => $newhexdump,
+          'typ' => $asn1_type,
+          'tlv_value' => $tlv_value,
+    ];
       if($remain == '') { // if remains string was empty & contents also empty, function return FALSE
         $stop = true;
       } else {
@@ -107,8 +108,8 @@ class asn1 {
    * @param int $maxDepth maximum parsing depth
    * @return array asn.1 structure recursively to specific depth
    */
-  public static function parse($hex, $maxDepth=5) {
-    $result = array();
+  public static function parse($hex, $maxDepth = 5): array {
+    $result = [];
     static $currentDepth = 0;
     if($asn1parse_array = self::oneParse($hex)) {
       foreach($asn1parse_array as $ff){
@@ -116,8 +117,8 @@ class asn1 {
         unset($info);
         $k = $ff['typ'];
         $v = $ff['tlv_value'];
-        $info['depth']=$currentDepth;
-        $info['hexdump']=$ff['newhexdump'];
+        $info['depth'] = $currentDepth;
+        $info['hexdump'] = $ff['newhexdump'];
         $info['type'] = $k;
         $info['typeName'] = self::type($k);
         $info['value_hex'] = $v;
@@ -125,7 +126,7 @@ class asn1 {
           if($k == '06') {
 
           } else if(in_array($k, ['13', '18'])) {
-            $info['value'] = hex2bin($info['value_hex']);
+            $info['value'] = hex2bin((string) $info['value_hex']);
           } else if(in_array($k, ['03', '02', 'a04'])) {
             $info['value'] = $v;
           } else {
@@ -145,7 +146,6 @@ class asn1 {
   }
   // =====End ASN.1 Parser section=====
 
-
   // =====Begin ASN.1 Builder section=====
   /**
    * create asn.1 TLV tag length, length length and value length
@@ -153,15 +153,15 @@ class asn1 {
    * @param string $str string value of asn.1
    * @return string hex of asn.1 TLV tag length
    */
-  protected static function asn1_header($str) {
-    $len = strlen($str)/2;
+  protected static function asn1_header($str): string {
+    $len = strlen($str) / 2;
     $ret = dechex($len);
-    if(strlen($ret)%2 != 0) {
+    if(strlen($ret) % 2 != 0) {
       $ret = "0$ret";
     }
-    $headerLength = strlen($ret)/2;
+    $headerLength = strlen($ret) / 2;
     if($len > 127) {
-      $ret = "8".$headerLength.$ret;
+      $ret = "8" . $headerLength . $ret;
     }
     return $ret;
   }
@@ -169,25 +169,25 @@ class asn1 {
   /**
    * create various dynamic function for asn1
    */
-  private static function asn1Tag($name) {
-    $functionList = array(
-      'seq'=>'30',
-      'oct'=>'04',
-      'obj'=>'06',
-      'bit'=>'03',
-      'printable'=>'13',
-      'int'=>'02',
-      'set'=>'31',
-      'expl'=>'a',
-      'utime'=>'17',
-      'gtime'=>'18',
-      'utf8'=>'0c',
-      'ia5'=>'16',
-      'visible'=>'1a',
-      't61'=>'14',
-      'impl'=>'80',
-      'other'=>''
-    );
+  private static function asn1Tag($name): string|false {
+    $functionList = [
+        'seq' => '30',
+        'oct' => '04',
+        'obj' => '06',
+        'bit' => '03',
+        'printable' => '13',
+        'int' => '02',
+        'set' => '31',
+        'expl' => 'a',
+        'utime' => '17',
+        'gtime' => '18',
+        'utf8' => '0c',
+        'ia5' => '16',
+        'visible' => '1a',
+        't61' => '14',
+        'impl' => '80',
+        'other' => '',
+    ];
     if(array_key_exists($name, $functionList)) {
       return $functionList[$name];
     } else {
@@ -197,26 +197,26 @@ class asn1 {
   }
 
   public static function __callStatic($func, $params) {
-    $func = strtolower($func);
+    $func = strtolower((string) $func);
       $asn1Tag = self::asn1Tag($func);
       if($asn1Tag !== false){
         $num = $asn1Tag; //valu of array
         $hex = $params[0];
         $val = $hex;
         if(in_array($func, ['printable', 'utf8', 'ia5', 'visible', 't61'])) { // ($string)
-          $val = bin2hex($hex);
+          $val = bin2hex((string) $hex);
         }
         if($func == 'int') {
-          $val = (strlen($val)%2 != 0)?"0$val":"$val";
+          $val = (strlen((string) $val) % 2 != 0) ? "0$val" : "$val";
         }
         if($func == 'expl') { //expl($num, $hex)
-          $num = $num.$params[0];
+          $num = $num . $params[0];
           $val = $params[1];
         }
         if($func == 'impl') { //impl($num="0")
-          $val = (!$val)?"00":$val;
-          $val = (strlen($val)%2 != 0)?"0$val":$val;
-          return $num.$val;
+          $val = (! $val) ? "00" : $val;
+          $val = (strlen((string) $val) % 2 != 0) ? "0$val" : $val;
+          return $num . $val;
         }
         if($func == 'other') { //OTHER($id, $hex, $chr = false)
           $id = $params[0];
@@ -224,9 +224,9 @@ class asn1 {
           $chr = @$params[2];
           $str = $hex;
           if($chr != false) {
-            $str = bin2hex($hex);
+            $str = bin2hex((string) $hex);
           }
-          $ret = "$id".self::asn1_header($str).$str;
+          $ret = "$id" . self::asn1_header($str) . $str;
           return $ret;
         }
         if($func == 'utime') {
@@ -235,10 +235,10 @@ class asn1 {
           date_default_timezone_set("UTC");
           $time = date("ymdHis", $time);
           date_default_timezone_set($oldTz);
-          $val = bin2hex($time."Z");
+          $val = bin2hex($time . "Z");
         }
         if($func == 'gtime') {
-          if(!$time = strtotime($params[0])) {
+          if(! $time = strtotime((string) $params[0])) {
              // echo "asn1::GTIME function strtotime cant recognize time!! please check at input=\"{$params[0]}\"";
             return false;
           }
@@ -246,10 +246,10 @@ class asn1 {
           // date_default_timezone_set("UTC");
           $time = date("YmdHis", $time);
           date_default_timezone_set($oldTz);
-          $val = bin2hex($time."Z");
+          $val = bin2hex($time . "Z");
         }
         $hdr = self::asn1_header($val);
-        return $num.$hdr.$val;
+        return $num . $hdr . $val;
     } else {
       // echo "asn1 \"$func\" not exists!";
     }
