@@ -21,15 +21,15 @@
 
 namespace ddn\sapp;
 
+use \ArrayAccess;
+use ddn\sapp\helpers\Buffer;
+use ddn\sapp\helpers\LoadHelpers;
 use ddn\sapp\pdfvalue\PDFValueObject;
 use ddn\sapp\pdfvalue\PDFValueSimple;
-use \ArrayAccess;
-
-use ddn\sapp\helpers\Buffer;
-
+use ReturnTypeWillChange;
 // Loading the functions
-use ddn\sapp\helpers\LoadHelpers;
-if (!defined("ddn\\sapp\\helpers\\LoadHelpers"))
+use Stringable;
+if (! defined("ddn\\sapp\\helpers\\LoadHelpers"))
     new LoadHelpers;
 
 use function ddn\sapp\helpers\p_debug;
@@ -38,24 +38,30 @@ use function ddn\sapp\helpers\p_error;
 use function ddn\sapp\helpers\p_warning;
 
 // The character used to end lines
-if (!defined('__EOL'))
+if (! defined('__EOL'))
     define('__EOL', "\n");
 
 /**
- * Class to gather the information of a PDF object: the OID, the definition and the stream. The purpose is to 
+ * Class to gather the information of a PDF object: the OID, the definition and the stream. The purpose is to
  *   ease the generation of the PDF entries for an individual object.
  */
-class PDFObject implements ArrayAccess {
+class PDFObject implements ArrayAccess, Stringable
+{
     protected static $_revisions;
+
     protected static $_xref_table_version;
 
-    protected $_oid = null;
     protected $_stream = null;
+
     protected $_value = null;
+
     protected $_generation;
 
-    
-    public function __construct($oid, $value = null, $generation = 0) {
+    public function __construct(
+        protected $_oid,
+        $value = null,
+        $generation = 0
+    ) {
         if ($generation !== 0)
             p_warning("Objects of non-zero generation are not fully checked... please double check your document and (if possible) please send examples via issues to https://github.com/dealfonso/sapp/issues/");
 
@@ -71,8 +77,6 @@ class PDFObject implements ArrayAccess {
             }
             $value = $obj;
         }
-
-        $this->_oid = $oid;
         $this->_value = $value;
         $this->_generation = $generation;
     }
@@ -81,7 +85,7 @@ class PDFObject implements ArrayAccess {
         return $this->_value->get_keys();
     }
 
-    public function set_oid($oid) {
+    public function set_oid($oid): void {
         $this->_oid = $oid;
     }
 
@@ -89,16 +93,18 @@ class PDFObject implements ArrayAccess {
         return $this->_generation;
     }
 
-    public function __toString() {
-        return  "$this->_oid 0 obj\n" .
+    public function __toString(): string {
+        return "$this->_oid 0 obj\n" .
             "$this->_value\n" .
-            ($this->_stream === null?"":
+            (
+                $this->_stream === null ? "" :
                 "stream\n" .
-                '...' . 
+                '...' .
                 "\nendstream\n"
             ) .
             "endobj\n";
     }
+
     /**
      * Converts the object to a well-formed PDF entry with a form like
      *  1 0 obj
@@ -109,16 +115,18 @@ class PDFObject implements ArrayAccess {
      *  endobj
      * @return pdfentry a string that contains the PDF entry
      */
-    public function to_pdf_entry() {
-        return  "$this->_oid 0 obj" . __EOL .
+    public function to_pdf_entry(): string {
+        return "$this->_oid 0 obj" . __EOL .
                 "$this->_value" . __EOL .
-                ($this->_stream === null?"":
+                (
+                    $this->_stream === null ? "" :
                     "stream\r\n" .
-                    $this->_stream . 
+                    $this->_stream .
                     __EOL . "endstream" . __EOL
                 ) .
                 "endobj" . __EOL;
     }
+
     /**
      * Gets the object ID
      * @return oid the object id
@@ -126,6 +134,7 @@ class PDFObject implements ArrayAccess {
     public function get_oid() {
         return $this->_oid;
     }
+
     /**
      * Gets the definition of the object (a PDFValue object)
      * @return value the definition of the object
@@ -133,7 +142,8 @@ class PDFObject implements ArrayAccess {
     public function get_value() {
         return $this->_value;
     }
-    protected static function FlateDecode($_stream, $params) {
+
+    protected static function FlateDecode($_stream, array $params) {
         switch ($params["Predictor"]->get_int()) {
             case 1:
                     return $_stream;
@@ -166,7 +176,7 @@ class PDFObject implements ArrayAccess {
         $columns = $params['Columns']->get_int();
 
         $row_len = $columns + 1;
-        $stream_len = strlen($_stream);
+        $stream_len = strlen((string) $_stream);
 
         // The previous row is zero
         $data_prev = str_pad("", $columns, chr(0));
@@ -177,7 +187,7 @@ class PDFObject implements ArrayAccess {
             $filter_byte = ord($_stream[$pos_i++]);
 
             // Get the current row
-            $data = substr($_stream, $pos_i, $columns);
+            $data = substr((string) $_stream, $pos_i, $columns);
             $pos_i += strlen($data);
 
             // Zero pad, in case that the content is not paired
@@ -185,18 +195,18 @@ class PDFObject implements ArrayAccess {
 
             // Depending on the filter byte of the row, we should unpack on one way or another
             switch ($filter_byte) {
-                case 0: 
+                case 0:
                     break;
-                case 1: 
+                case 1:
                     for ($i = 1; $i < $columns; $i++)
-                        $data[$i] = ($data[$i] + $data[$i-1]) % 256;
+                        $data[$i] = ($data[$i] + $data[$i - 1]) % 256;
                     break;
-                case 2: 
+                case 2:
                     for ($i = 0; $i < $columns; $i++) {
                         $data[$i] = chr((ord($data[$i]) + ord($data_prev[$i])) % 256);
                     }
                     break;
-                default: 
+                default:
                     return p_error("Unsupported stream");
             }
 
@@ -219,15 +229,15 @@ class PDFObject implements ArrayAccess {
         if (isset($this->_value['Filter'])) {
             switch ($this->_value['Filter']) {
                 case '/FlateDecode':
-                    $DecodeParams = $this->_value['DecodeParms']??[];
+                    $DecodeParams = $this->_value['DecodeParms'] ?? [];
                     $params = [
-                        "Columns" => $DecodeParams['Columns']??new PDFValueSimple(0),
-                        "Predictor" => $DecodeParams['Predictor']??new PDFValueSimple(1),
-                        "BitsPerComponent" => $DecodeParams['BitsPerComponent']??new PDFValueSimple(8),
-                        "Colors" => $DecodeParams['Colors']??new PDFValueSimple(1)
+                        "Columns" => $DecodeParams['Columns'] ?? new PDFValueSimple(0),
+                        "Predictor" => $DecodeParams['Predictor'] ?? new PDFValueSimple(1),
+                        "BitsPerComponent" => $DecodeParams['BitsPerComponent'] ?? new PDFValueSimple(8),
+                        "Colors" => $DecodeParams['Colors'] ?? new PDFValueSimple(1),
                     ];
                     return self::FlateDecode(gzuncompress($this->_stream), $params);
-                    
+
                     break;
                 default:
                     return p_error('unknown compression method ' . $this->_value['Filter']);
@@ -235,11 +245,12 @@ class PDFObject implements ArrayAccess {
         }
         return $this->_stream;
     }
+
     /**
      * Sets the stream for the object (overwrites a previous existing stream)
      * @param stream the stream for the object
      */
-    public function set_stream($stream, $raw = true) {
+    public function set_stream($stream, $raw = true): void {
         if ($raw === true) {
             $this->_stream = $stream;
             return;
@@ -247,55 +258,58 @@ class PDFObject implements ArrayAccess {
         if (isset($this->_value['Filter'])) {
             switch ($this->_value['Filter']) {
                 case '/FlateDecode':
-                    $stream = gzcompress($stream);
+                    $stream = gzcompress((string) $stream);
                     break;
                 default:
                     p_error('unknown compression method ' . $this->_value['Filter']);
             }
         }
-        $this->_value['Length'] = strlen($stream);
+        $this->_value['Length'] = strlen((string) $stream);
         $this->_stream = $stream;
-    }    
+    }
+
     /**
      * The next functions enble to make use of this object in an array-like manner,
      *  using the name of the fields as positions in the array. It is useful is the
      *  value is of type PDFValueObject or PDFValueList, using indexes
      */
-
-    /** 
+    /**
      * Sets the value of the field offset, using notation $obj['field'] = $value
      * @param field the field to set the value
      * @param value the value to set
      * @return void
      */
-    public function offsetSet($field, $value) : void {
+    public function offsetSet($field, $value): void {
         $this->_value[$field] = $value;
     }
+
     /**
      * Checks whether the field exists in the object or not (or if the index exists
      *   in the list)
      * @param field the field to check wether exists or not
      * @return exists true if the field exists; false otherwise
      */
-    public function offsetExists ( $field ) : bool {
+    public function offsetExists ( $field ): bool {
         return $this->_value->offsetExists($field);
     }
+
     /**
      * Gets the value of the field (or the value at position)
      * @param field the field to get the value
      * @return value the value of the field
      */
-    #[\ReturnTypeWillChange]
-    public function offsetGet ( $field ) { 
+    #[ReturnTypeWillChange]
+    public function offsetGet ( $field ) {
         return $this->_value[$field];
     }
+
     /**
      * Unsets the value of the field (or the value at position)
      * @param field the field to unset the value
      */
-    public function offsetUnset($field ) : void {
+    public function offsetUnset($field ): void {
         $this->_value->offsetUnset($field);
-    }    
+    }
 
     public function push($v) {
         return $this->_value->push($v);
