@@ -1,25 +1,27 @@
 <?php
+
 namespace ddn\sapp\helpers;
 
 class UUID {
-  public static function v3($namespace, $name) {
-    if(!self::is_valid($namespace)) return false;
+  public static function v3($namespace, string $name): false|string {
+    if(! self::is_valid($namespace)) return false;
 
     // Get hexadecimal components of namespace
-    $nhex = str_replace(array('-','{','}'), '', $namespace);
+    $nhex = str_replace(['-', '{', '}'], '', $namespace);
 
     // Binary Value
     $nstr = '';
 
     // Convert Namespace UUID to bits
-    for($i = 0; $i < strlen($nhex); $i+=2) {
-      $nstr .= chr(hexdec($nhex[$i].$nhex[$i+1]));
+    for($i = 0; $i < strlen($nhex); $i += 2) {
+      $nstr .= chr(hexdec($nhex[$i] . $nhex[$i + 1]));
     }
 
     // Calculate hash value
     $hash = md5($nstr . $name);
 
-    return sprintf('%08s-%04s-%04x-%04x-%12s',
+    return sprintf(
+        '%08s-%04s-%04x-%04x-%12s',
 
       // 32 bits for "time_low"
       substr($hash, 0, 8),
@@ -41,11 +43,13 @@ class UUID {
     );
   }
 
-  public static function v4() {
-    return sprintf('%04x%04x-%04x-%04x-%04x-%04x%04x%04x',
+  public static function v4(): string {
+    return sprintf(
+        '%04x%04x-%04x-%04x-%04x-%04x%04x%04x',
 
       // 32 bits for "time_low"
-      mt_rand(0, 0xffff), mt_rand(0, 0xffff),
+      mt_rand(0, 0xffff),
+        mt_rand(0, 0xffff),
 
       // 16 bits for "time_mid"
       mt_rand(0, 0xffff),
@@ -60,28 +64,31 @@ class UUID {
       mt_rand(0, 0x3fff) | 0x8000,
 
       // 48 bits for "node"
-      mt_rand(0, 0xffff), mt_rand(0, 0xffff), mt_rand(0, 0xffff)
+      mt_rand(0, 0xffff),
+        mt_rand(0, 0xffff),
+        mt_rand(0, 0xffff)
     );
   }
 
-  public static function v5($namespace, $name) {
-    if(!self::is_valid($namespace)) return false;
+  public static function v5($namespace, string $name): false|string {
+    if(! self::is_valid($namespace)) return false;
 
     // Get hexadecimal components of namespace
-    $nhex = str_replace(array('-','{','}'), '', $namespace);
+    $nhex = str_replace(['-', '{', '}'], '', $namespace);
 
     // Binary Value
     $nstr = '';
 
     // Convert Namespace UUID to bits
-    for($i = 0; $i < strlen($nhex); $i+=2) {
-      $nstr .= chr(hexdec($nhex[$i].$nhex[$i+1]));
+    for($i = 0; $i < strlen($nhex); $i += 2) {
+      $nstr .= chr(hexdec($nhex[$i] . $nhex[$i + 1]));
     }
 
     // Calculate hash value
     $hash = sha1($nstr . $name);
 
-    return sprintf('%08s-%04s-%04x-%04x-%12s',
+    return sprintf(
+        '%08s-%04s-%04x-%04x-%12s',
 
       // 32 bits for "time_low"
       substr($hash, 0, 8),
@@ -103,8 +110,8 @@ class UUID {
     );
   }
 
-  public static function is_valid($uuid) {
-    return preg_match('/^\{?[0-9a-f]{8}\-?[0-9a-f]{4}\-?[0-9a-f]{4}\-?'.
-                      '[0-9a-f]{4}\-?[0-9a-f]{12}\}?$/i', $uuid) === 1;
+  public static function is_valid($uuid): bool {
+    return preg_match('/^\{?[0-9a-f]{8}\-?[0-9a-f]{4}\-?[0-9a-f]{4}\-?' .
+                      '[0-9a-f]{4}\-?[0-9a-f]{12}\}?$/i', (string) $uuid) === 1;
   }
 }
