@@ -21,33 +21,24 @@
 
 namespace ddn\sapp\helpers;
 
-use ddn\sapp\PDFBaseDoc;
-use ddn\sapp\PDFBaseObject;
-use ddn\sapp\pdfvalue\PDFValueObject;
 use ddn\sapp\pdfvalue\PDFValueList;
+use ddn\sapp\pdfvalue\PDFValueObject;
 use ddn\sapp\pdfvalue\PDFValueReference;
 use ddn\sapp\pdfvalue\PDFValueType;
-use ddn\sapp\pdfvalue\PDFValueHexString;
-use ddn\sapp\pdfvalue\PDFValueString;
+use finfo;
 
-use function ddn\sapp\helpers\get_random_string;
-use function ddn\sapp\helpers\mime_to_ext;
-use function ddn\sapp\helpers\_parsejpg;
-use function ddn\sapp\helpers\_parsepng;
-use function ddn\sapp\helpers\p_error;
-
-function tx($x, $y) {
+function tx($x, $y): string {
     return sprintf(" 1 0 0 1 %.2F %.2F cm", $x, $y);
 }
-function sx($w, $h) {
+function sx($w, $h): string {
     return sprintf(" %.2F 0 0 %.2F 0 0 cm", $w, $h);
 }
-function deg2rad($angle) {
+function deg2rad($angle): float {
     return $angle * pi() / 180;
 }
-function rx($angle) {
+function rx($angle): string {
     $angle = deg2rad($angle);
-    return sprintf(" %.2F %.2F %.2F %.2F 0 0 cm", cos($angle), sin($angle), -sin($angle), cos($angle)); 
+    return sprintf(" %.2F %.2F %.2F %.2F 0 0 cm", cos($angle), sin($angle), -sin($angle), cos($angle));
 }
 
 /**
@@ -55,24 +46,25 @@ function rx($angle) {
  *   NOTE: the image inclusion is taken from http://www.fpdf.org/; this is a translation
  *         of function _putimage
  */
-function _create_image_objects($info, $object_factory) { 
+function _create_image_objects($info, $object_factory): array {
     $objects = [];
 
-    $image = call_user_func($object_factory,
+    $image = call_user_func(
+        $object_factory,
         [
             'Type' => '/XObject',
             'Subtype' => '/Image',
             'Width' => $info['w'],
             'Height' => $info['h'],
-            'ColorSpace' => [ ],
+            'ColorSpace' => [],
             'BitsPerComponent' => $info['bpc'],
-            'Length' => strlen($info['data']),
-        ]            
+            'Length' => strlen((string) $info['data']),
+        ]
     );
 
     switch ($info['cs']) {
         case 'Indexed':
-            $data = gzcompress($info['pal']);
+            $data = gzcompress((string) $info['pal']);
             $streamobject = call_user_func($object_factory, [
                 'Filter' => '/FlateDecode',
                 'Length' => strlen($data),
@@ -80,7 +72,7 @@ function _create_image_objects($info, $object_factory) {
             $streamobject->set_stream($data);
 
             $image['ColorSpace']->push([
-                '/Indexed', '/DeviceRGB', (strlen($info['pal']) / 3) - 1, new PDFValueReference($streamobject->get_oid())
+                '/Indexed', '/DeviceRGB', (strlen((string) $info['pal']) / 3) - 1, new PDFValueReference($streamobject->get_oid()),
             ]);
             array_push($objects, $streamobject);
             break;
@@ -102,13 +94,13 @@ function _create_image_objects($info, $object_factory) {
 
     if (isset($info['smask'])) {
         $smaskinfo = [
-            'w' => $info['w'], 
-            'h' => $info['h'], 
-            'cs' => 'DeviceGray', 
-            'bpc' => 8, 
-            'f' => $info['f'], 
-            'dp' => '/Predictor 15 /Colors 1 /BitsPerComponent 8 /Columns '.$info['w'],
-            'data' => $info['smask']
+            'w' => $info['w'],
+            'h' => $info['h'],
+            'cs' => 'DeviceGray',
+            'bpc' => 8,
+            'f' => $info['f'],
+            'dp' => '/Predictor 15 /Colors 1 /BitsPerComponent 8 /Columns ' . $info['w'],
+            'data' => $info['smask'],
         ];
 
         // In principle, it may return multiple objects
@@ -124,12 +116,12 @@ function _create_image_objects($info, $object_factory) {
     return $objects;
 }
 
-function is_base64($string){
+function is_base64($string): bool{
     // Check if there are valid base64 characters
-    if (!preg_match('/^[a-zA-Z0-9\/\r\n+]*={0,2}$/', $string)) return false;
+    if (! preg_match('/^[a-zA-Z0-9\/\r\n+]*={0,2}$/', (string) $string)) return false;
 
     // Decode the string in strict mode and check the results
-    $decoded = base64_decode($string, true);
+    $decoded = base64_decode((string) $string, true);
     if(false === $decoded) return false;
 
     // Encode the string again
@@ -149,22 +141,22 @@ function is_base64($string){
  * @param w width of the rectangle in which to appear the image (image will be scaled, and the units are "content-defined" (i.e. depending on the size of the page))
  * @param h height of the rectangle in which to appear the image (image will be scaled, and the units are "content-defined" (i.e. depending on the size of the page))
  * @param angle the rotation angle in degrees; the image will be rotated using the center
- * @param keep_proportions if true, the image will keep the proportions when rotated, then the image will not occupy the full 
+ * @param keep_proportions if true, the image will keep the proportions when rotated, then the image will not occupy the full
  * @return result an array with the next fields:
  *                  "images": objects of the corresponding images (i.e. position [0] is the image, the rest elements are masks, if needed)
  *                  "resources": PDFValueObject with keys that needs to be incorporated to the resources of the object in which the images will appear
  *                  "alpha": true if the image has alpha
  *                  "command": pdf command to draw the image
  */
-function _add_image($object_factory, $filename, $x=0, $y=0, $w=0, $h=0, $angle = 0, $keep_proportions = true) {
+function _add_image($object_factory, $filename, $x = 0, $y = 0, $w = 0, $h = 0, $angle = 0, $keep_proportions = true) {
 
     if (empty($filename))
         return p_error('invalid image name or stream');
 
     if ($filename[0] === '@') {
-        $filecontent = substr($filename, 1);
+        $filecontent = substr((string) $filename, 1);
     } else if (is_base64($filename)) {
-        $filecontent = base64_decode($filename);
+        $filecontent = base64_decode((string) $filename, true);
     } else {
         $filecontent = @file_get_contents($filename);
 
@@ -172,7 +164,7 @@ function _add_image($object_factory, $filename, $x=0, $y=0, $w=0, $h=0, $angle =
             return p_error("failed to get the image");
     }
 
-    $finfo = new \finfo();
+    $finfo = new finfo();
     $content_type = $finfo->buffer($filecontent, FILEINFO_MIME_TYPE);
 
     $ext = mime_to_ext($content_type);
@@ -200,14 +192,14 @@ function _add_image($object_factory, $filename, $x=0, $y=0, $w=0, $h=0, $angle =
     if ($h === null)
         $h = -96;
 
-    if($w<0)
-        $w = -$info['w']*72/$w;
-    if($h<0)
-        $h = -$info['h']*72/$h;
-    if($w==0)
-        $w = $h*$info['w']/$info['h'];
-    if($h==0)
-        $h = $w*$info['h']/$info['w'];
+    if($w < 0)
+        $w = -$info['w'] * 72 / $w;
+    if($h < 0)
+        $h = -$info['h'] * 72 / $h;
+    if($w == 0)
+        $w = $h * $info['w'] / $info['h'];
+    if($h == 0)
+        $h = $w * $info['h'] / $info['w'];
 
     $images_objects = _create_image_objects($info, $object_factory);
 
@@ -253,16 +245,21 @@ function _add_image($object_factory, $filename, $x=0, $y=0, $w=0, $h=0, $angle =
     if ($angle != 0) {
         $data .= tx(0.5, 0.5);
         $data .= rx($angle);
-        $data .= tx(-0.5,-0.5);
+        $data .= tx(-0.5, -0.5);
     }
     $data .= sprintf(" /%s Do Q", $info['i']);
 
     $resources = new PDFValueObject( [
-        'ProcSet' => [ '/PDF', '/Text', '/ImageB', '/ImageC', '/ImageI' ],
+        'ProcSet' => ['/PDF', '/Text', '/ImageB', '/ImageC', '/ImageI'],
         'XObject' => new PDFValueObject ([
-            $info['i'] => new PDFValueReference($images_objects[0]->get_oid()),                        
-        ])
+            $info['i'] => new PDFValueReference($images_objects[0]->get_oid()),
+        ]),
     ]);
 
-    return [ "image" => $images_objects[0], 'command' => $data, 'resources' => $resources, 'alpha' => $add_alpha ];
+    return [
+        "image" => $images_objects[0],
+        'command' => $data,
+        'resources' => $resources,
+        'alpha' => $add_alpha,
+    ];
 }
