@@ -22,7 +22,7 @@
 namespace ddn\sapp\pdfvalue;
 
 class PDFValueString extends PDFValue {
-    public function __toString() {
+    public function __toString(): string {
         return "(" . $this->value . ")";
     }
 }
